requirements.txt
streamlit==1.25.0
google-cloud-retail==1.16.2
google-cloud-aiplatform==1.28.1
